Move TargetConfig to TargetEntity cpp/h
