+++ source/TargetEntity.h
@@ -1,29 +1,16 @@
 #pragma once
 #include <G3D/G3D.h>
 
-
-class TargetConfig;
-
-//#define DRAW_BOUNDING_SPHERES	1		// Uncomment this to draw bounding spheres (useful for target sizing)
-#define BOUNDING_SPHERE_RADIUS	0.5		///< Use a 0.5m radius for sizing here
-
-// 1 - 2 ^ (1/3) makes multiples of 2 regular
-#define TARGET_MODEL_ARRAY_SCALING 0.25992104989f
-// Model size offset
-#define TARGET_MODEL_ARRAY_OFFSET 40.0f
-// Number of target model sizes
-#define TARGET_MODEL_SCALE_COUNT 50
-
-struct Destination{
+struct Destination {
 public:
-	Point3 position = Point3(0,0,0);
+	Point3 position = Point3(0, 0, 0);
 	SimTime time = 0.0;
 
 	Destination() {
 		position = Point3(0, 0, 0);
 		time = 0.0;
 	}
-	
+
 	Destination(Point3 pos, SimTime t) {
 		position = pos;
 		time = t;
@@ -56,6 +43,69 @@ public:
 		return HashTrait<Point3>::hashCode(position) ^ (int)time;
 	}
 };
+
+/** Class for representing a given target configuration */
+class TargetConfig : public ReferenceCountedObject {
+public:
+	String			id;										///< Trial ID to indentify affiliated trial runs
+	//bool    		elevLocked = false;						///< Elevation locking
+	bool			upperHemisphereOnly = false;            ///< Limit flying motion to upper hemisphere only
+	bool			logTargetTrajectory = true;				///< Log this target's trajectory
+	Array<float>	distance = { 30.0f, 40.0f };			///< Distance to the target
+	Array<float>	motionChangePeriod = { 1.0f, 1.0f };	///< Range of motion change period in seconds
+	Array<float>	speed = { 0.0f, 5.5f };					///< Range of angular velocities for target
+	Array<float>	eccH = { 5.0f, 15.0f };					///< Range of initial horizontal eccentricity
+	Array<float>	eccV = { 0.0f, 2.0f };					///< Range of initial vertical eccentricity
+	Array<float>	size = { 0.2f, 0.2f };					///< Visual size of the target (in degrees)
+	bool			jumpEnabled = false;					///< Flag indicating whether the target jumps
+	Array<float>	jumpPeriod = { 2.0f, 2.0f };			///< Range of time period between jumps in seconds
+	Array<float>	jumpSpeed = { 2.0f, 5.5f };				///< Range of jump speeds in meters/s
+	Array<float>	accelGravity = { 9.8f, 9.8f };			///< Range of acceleration due to gravity in meters/s^2
+	Array<Destination> destinations;						///< Array of destinations to traverse
+	String			destSpace = "world";					///< Space to use for destinations (implies offset) can be "world" or "player"
+	int				respawnCount = 0;						///< Number of times to respawn
+	AABox			spawnBounds;							///< Spawn position bounding box
+	AABox			moveBounds;								///< Movemvent bounding box
+	Array<bool>		axisLock = { false, false, false };		///< Array of axis lock values
+
+	String			destroyDecal = "explosion_01.png";		///< Decal to use for destroy event
+	float			destroyDecalScale = 1.0;				///< Scale to apply to destroy (relative to target size)
+	RealTime		destroyDecalDuration = 0.1;				///< Destroy decal display duration
+
+	String			hitSound = "sound/fpsci_ding_100ms.wav";///< Sound to play when target is hit (but not destoyed)
+	float			hitSoundVol = 1.0f;						///< Hit sound volume
+	String          destroyedSound = "sound/fpsci_destroy_150ms.wav";		///< Sound to play when target destroyed
+	float           destroyedSoundVol = 1.0f;
+
+	Any modelSpec = PARSE_ANY(ArticulatedModel::Specification{			///< Basic model spec for target
+		filename = "model/target/target.obj";
+		cleanGeometrySettings = ArticulatedModel::CleanGeometrySettings{
+					allowVertexMerging = true;
+					forceComputeNormals = false;
+					forceComputeTangents = false;
+					forceVertexMerging = true;
+					maxEdgeLength = inf;
+					maxNormalWeldAngleDegrees = 0;
+					maxSmoothAngleDegrees = 0;
+			};
+		});
+
+	TargetConfig() {}
+	TargetConfig(const Any& any);
+
+	static TargetConfig load(const String& filename);
+	Any toAny(const bool forceAll = false) const;
+};
+
+//#define DRAW_BOUNDING_SPHERES	1		// Uncomment this to draw bounding spheres (useful for target sizing)
+#define BOUNDING_SPHERE_RADIUS	0.5		///< Use a 0.5m radius for sizing here
+
+// 1 - 2 ^ (1/3) makes multiples of 2 regular
+#define TARGET_MODEL_ARRAY_SCALING 0.25992104989f
+// Model size offset
+#define TARGET_MODEL_ARRAY_OFFSET 40.0f
+// Number of target model sizes
+#define TARGET_MODEL_SCALE_COUNT 50
 
 class TargetEntity : public VisibleEntity {
 protected:
